+++ src/Administration/Resources/app/administration/test/e2e/cypress/integration/general/sw-dashboard/read.spec.js
@@ -45,10 +45,7 @@ describe('Dashboard: Test first sight of the Administration', () => {
 
     it('@base @general: read dashboard', () => {
         // Check today stats
-        cy.get('.sw-dashboard-index__intro-content').contains('Welcome to Shopware 6');
-        cy.get('.sw-dashboard-index__intro-image img')
-            .should('have.attr', 'src')
-            .should('include', '/bundles/administration/static/img/dashboard-logo.svg');
+        cy.get('.sw-dashboard-index__welcome-title').contains('Welcome to Shopware');
         cy.get('.sw-dashboard-index__intro-stats-today-single-stat-number-value').should('be.visible');
         cy.get('.sw-dashboard-index__intro-stats-today-single-stat-number-value').contains('1');
         cy.get('.sw-dashboard-index__intro-stats-today-single-stat-number-value').contains(product.price[0].gross);