+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This is the official changelog index of Shopware 6. Here you find a registry of 
 *  [NEXT-14735 - Fix image slider navigation of CMS product detail page on mobile](/changelog/release-6-4-0-0/2021-04-12-fix-image-slider-navigation-of-cms-product-detail-page-on-mobile.md)
 *  [NEXT-14548 - Fix missing close button of offcanvas on cms product page](/changelog/release-6-4-0-0/2021-04-01-fix-missing-close-button-of-offcanvas-on-cms-product-page.md)
 *  [NEXT-14415 - Set local time configuration for date range rule builder](/changelog/release-6-4-0-0/2021-03-29-set-local-time-configuration-for-date-range-rule-builder.md)
+*  [NEXT-14705 - Change Dashboard welcome screen](/changelog/release-6-4-0-0/2021-04-12-change-dashboard-welcome-screen.md)
 *  [NEXT-14348 - Fix change variant price on detail page](/changelog/release-6-4-0-0/2021-03-23-fix-change-variant-price-on-detail-page.md)
 *  [NEXT-11082 - Fix wrong time setting in datepicker component](/changelog/release-6-4-0-0/2021-03-15-fix-wrong-time-setting-in-datepicker-component.md)
 *  [NEXT-13789 - Allow null type in ProductConfiguratorSettingEntity mediaId](/changelog/release-6-4-0-0/2021-02-16-Allow-Null-Type-In-ProductConfiguratorSettingEntity-MediaId.md) ([Sebastian Diez](https://github.com/s-diez))